test: Move record tests to tests/ directory
